backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langchain==0.3.0
langchain-groq==0.2.0
langgraph==0.2.28
qdrant-client==1.11.3
fastembed==0.4.1
python-dotenv==1.0.1
playwright==1.47.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
httpx==0.27.2
python-jose==3.3.0
sse-starlette==2.1.3
